requirements.txt
streamlit
plotly
pandas
numpy
folium
matplotlib-inline
haversine
streamlit-folium
pillow